examples/12-event-selector/requirements.txt
soorma-sdk>=0.8.0
soorma-common>=0.8.0
python-dotenv>=1.0.0
